requirements.txt
google-generativeai==0.6.0
flask==3.0.3
waitress==3.0.0